split tiles
